Create Recipes Folder, Images Folder, and Recipe Page
Since there weren't any recipes on the website, I added a "recipes" folder for better organization.

I also added created an HTML file in that folder, which contains a recipe for chocolate truffle pie. On that file, there is an image of the chocolate truffle pie, a short description of the finished product, an ingredients list, and the recipe steps.

I also created an "images" folder to store the image of the chocolate truffle pie, and to store any future recipe images.

Signed-off-by: Nathan Olson <921778+Lutrome@users.noreply.github.com>